feat: added ranked mixt push and captains mix

+++ bot/src/interactions/commands/setup_mix_captains.ts
@@ -2,6 +2,7 @@ import { SlashCommandBuilder } from "@discordjs/builders";
 import { ChatInputCommandInteraction } from "discord.js";
 import { BOT_ADMIN_ROLE } from "../../constants";
 import { ICommandHandler } from "../../handlers/commandHandler";
+import { authorizationService } from "../../services/authorizationService";
 import { interactionUtils } from "../../services/interactionUtils";
 import { matchmakingService } from "../../services/matchmakingService";
 import { LINEUP_TYPE_CAPTAINS, LINEUP_VISIBILITY_TEAM, teamService } from "../../services/teamService";
@@ -15,8 +16,6 @@ export default {
                 .setRequired(true)
                 .setDescription('The number of players in each team')
                 .addChoices(
-                    { name: '1', value: 1 },
-                    { name: '2', value: 2 },
                     { name: '3', value: 3 },
                     { name: '4', value: 4 },
                     { name: '5', value: 5 },
@@ -27,7 +26,10 @@ export default {
                     { name: '10', value: 10 },
                     { name: '11', value: 11 }
                 )
-            ),
+            )
+            .addBooleanOption(option => option.setName('ranked')
+                .setRequired(false)
+                .setDescription('Indicates if this lineup is ranked and should update ratings')),
     authorizedRoles: [BOT_ADMIN_ROLE],
     async execute(interaction: ChatInputCommandInteraction) {
         const team = await teamService.findTeamByGuildId(interaction.guildId!)
@@ -36,8 +38,14 @@ export default {
             return
         }
 
+        const ranked = interaction.options.getBoolean("ranked") !== false
+        if (ranked && !authorizationService.isOfficialDiscord(interaction.guildId!)) {
+            await interaction.reply({ content: "⛔ Only official community discords can create ranked mixes", ephemeral: true })
+            return
+        }
+
         const lineupSize = interaction.options.getInteger("size")!
-        const lineup = teamService.createLineup(interaction.channelId, lineupSize, undefined, false, false, team, LINEUP_TYPE_CAPTAINS, LINEUP_VISIBILITY_TEAM)
+        const lineup = teamService.createLineup(interaction.channelId, lineupSize, undefined, false, ranked, team, LINEUP_TYPE_CAPTAINS, LINEUP_VISIBILITY_TEAM)
         await teamService.upsertLineup(lineup)
 
         await matchmakingService.deleteLineupQueuesByChannelId(interaction.channelId)
